// src/interfaces/FormInputs.ts
interface FormInputs {
  username: HTMLInputElement | null;
  about: HTMLInputElement | null;
  firstName: HTMLInputElement | null;
  lastName: HTMLInputElement | null;
  email: HTMLInputElement | null;
  country: HTMLInputElement | null;
  streetAddress: HTMLInputElement | null;
  city: HTMLInputElement | null;
  region: HTMLInputElement | null;
  postalCode: HTMLInputElement | null;
  age: HTMLInputElement | null;
  comments: HTMLInputElement | null;
  candidates: HTMLInputElement | null;
  offers: HTMLInputElement | null;
  pushEverything: HTMLInputElement | null;
  pushEmail: HTMLInputElement | null;
  pushNothing: HTMLInputElement | null;
}

export default FormInputs;